# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UNAnimity",
    "version": "0.2.4"
  },
  "servers": [
    {
      "url": "https://unanimity.be"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "Send user information and login credentials for a new account on the application",
        "requestBody": {
          "content": {
            "json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterInformation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request : the given format is incorrect",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "json format is required"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity : missing parameter",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "property 'email' is missing"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "Send login credentials for authentication",
        "requestBody": {
          "content": {
            "json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "The username credential"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password credential"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessUserInformation"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request : the given format is incorrect",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "json format is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "Wrong email/password association"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity : missing parameter",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "property 'email' is missing"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/activate": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "id",
                  "token"
                ],
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "The user id, it must be related to a valid user"
                  },
                  "token": {
                    "type": "string",
                    "description": "The generated token sent for the activation."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request : the given format is incorrect",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "json format is required"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "message": "reason: Incorrect activation scheme"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user_id"
      }
    },
    "schemas": {
      "Info": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "description": "The street name of the address of the user"
          },
          "number": {
            "type": "integer",
            "description": "The house number of the address of the user",
            "format": "int32"
          },
          "city": {
            "type": "string",
            "description": "The city where the user lives in"
          },
          "zipcode": {
            "type": "integer",
            "description": "The zip code of the address of the user",
            "format": "int32"
          },
          "country": {
            "type": "string",
            "description": "The country where the user lives in"
          },
          "box_number": {
            "type": "string",
            "description": "The mail box number of the user"
          }
        },
        "required": [
          "street",
          "number",
          "city",
          "zipcode",
          "country"
        ]
      },
      "UserRegisterInformation": {
        "type": "object",
        "required": [
          "email",
          "firstname",
          "lastname",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address for the user to be contacted"
          },
          "password": {
            "type": "string",
            "description": "The password that the user will use to log on the platform"
          },
          "firstname": {
            "type": "string",
            "description": "The first name of the user"
          },
          "lastname": {
            "type": "string",
            "description": "The last name of the user"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "phone": {
            "type": "string",
            "description": "The phone number of the user"
          }
        }
      },
      "AuthSuccessUserInformation": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "The id of the user"
              },
              "firstname": {
                "type": "string",
                "description": "The first name of the user"
              },
              "lastname": {
                "type": "string",
                "description": "The last name of the user"
              },
              "address": {
                "$ref": "#/components/schemas/Address"
              },
              "phone": {
                "type": "string",
                "description": "The phone number of the user"
              },
              "creation_date": {
                "type": "string",
                "description": "The date when the user was created"
              },
              "last_connection": {
                "type": "string",
                "description": "The date when the user connected last time"
              }
            }
          }
        }
      },
...
    }
  }
}